Moved audio analysis to src folder
